finish renaming, remove patch.py
