+++ ext/opentelemetry-ext-docker-tests/tests/redis/test_redis_functional.py
@@ -15,7 +15,7 @@
 import redis
 
 from opentelemetry import trace
-from opentelemetry.instrumentation.redis import RedisInstrumentor
+from opentelemetry.ext.redis import RedisInstrumentor
 from opentelemetry.test.test_base import TestBase
