+++ ext/opentelemetry-ext-redis/setup.cfg
@@ -55,4 +55,4 @@ where = src
 
 [options.entry_points]
 opentelemetry_instrumentor =
-    redis = opentelemetry.instrumentation.redis:RedisInstrumentor
+    redis = opentelemetry.ext.redis:RedisInstrumentor
