+++ ext/opentelemetry-ext-redis/src/opentelemetry/ext/redis/util.py
