+++ ext/opentelemetry-ext-redis/src/opentelemetry/ext/redis/version.py
